requirements.txt
# テストフレームワーク
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# 本体依存
transformers>=4.35.0
torch>=2.2.0
qdrant-client>=1.7.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-rust>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-c>=0.20.0
tree-sitter-cpp>=0.20.0
PyYAML>=6.0
pathspec>=0.11.0